Selectively suppress zero_extend simplification of sign_extend lshiftrt
Optimisation simplifies sign extend of logical right shift above 0
into a zero extend, as the sign bit will be zero, however this
optimisation is only better on platforms that naturally zero extend.
We add logic to skip this optimisation on platforms that naturally
sign extend the result of operations on narrower width modes.